# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/disbursements": {
      "post": {
        "description": "Send money to a customer Airtel Money wallet",
        "operationId": "DisbursementsController_createDisbursement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Disbursement request details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDisbursementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Disbursement created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisbursementResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (MSISDN format, amount, PIN format, etc.)"
          },
          "409": {
            "description": "Duplicate externalId (returns existing disbursement)"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a new disbursement",
        "tags": [
          "Disbursements"
        ]
      },
      "get": {
        "description": "Retrieve paginated list of disbursements with optional filtering",
        "operationId": "DisbursementsController_listDisbursements",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max: 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by disbursement status",
            "schema": {
              "enum": [
                "PENDING",
                "PROCESSING",
                "SUCCESS",
                "FAILED",
                "TIMEOUT",
                "BOUNCED",
                "REFUNDED"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation date (ISO 8601, inclusive)",
            "schema": {
              "example": "2024-02-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation date (ISO 8601, inclusive)",
            "schema": {
              "example": "2024-02-28T23:59:59Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DisbursementResponseDto"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters or validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List disbursements",
        "tags": [
          "Disbursements"
        ]
      }
    },
    "/api/v1/disbursements/{id}": {
      "get": {
        "description": "Retrieve details of a specific disbursement",
        "operationId": "DisbursementsController_getDisbursement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Disbursement UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disbursement found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisbursementResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters or validation failed"
          },
          "404": {
            "description": "Disbursement not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get disbursement by ID",
        "tags": [
          "Disbursements"
        ]
      }
    },
    "/api/v1/disbursements/portal": {
      "get": {
        "operationId": "DisbursementsController_listDisbursementsFromPortal",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by disbursement status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PROCESSING",
                "SUCCESS",
                "FAILED",
                "TIMEOUT",
                "BOUNCED",
                "REFUNDED",
                "REFUND_PROCESSING",
                "REFUND_FAILED"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for filtering (ISO 8601)",
            "schema": {
              "example": "2024-02-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for filtering (ISO 8601)",
            "schema": {
              "example": "2024-02-06T23:59:59Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disbursements fetched from admin portal"
          },
          "400": {
            "description": "Invalid request parameters or validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "List disbursements (portal)",
        "tags": [
          "Disbursements"
        ]
      }
    },
...
  },
  "info": {
    "title": "Payment Gateway",
    "description": "API documentation for Payment Gateway",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "x-tenant-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id"
      }
    },
    "schemas": {
...
      "CreateDisbursementDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "description": "Payment provider (AIRTEL or MTN)",
            "enum": [
              "AIRTEL",
              "MTN"
            ],
            "example": "AIRTEL"
          },
          "externalId": {
            "type": "string",
            "description": "Client-provided idempotency key for duplicate prevention",
            "example": "order-2024-001-disbursement",
            "maxLength": 255
          },
          "payeeMsisdn": {
            "type": "string",
            "description": "Recipient mobile number without country code (e.g., 0977123456)",
            "example": "0977123456",
            "minLength": 9,
            "maxLength": 15
          },
          "amount": {
            "type": "number",
            "description": "Disbursement amount in specified currency",
            "example": 500.5,
            "minimum": 0.01
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217 currency code",
            "example": "ZMW",
            "default": "ZMW",
            "minLength": 3,
            "maxLength": 3
          },
          "reference": {
            "type": "string",
            "description": "Business reference for reconciliation",
            "example": "INV-2024-001",
            "maxLength": 255
          },
          "pin": {
            "type": "string",
            "description": "4-digit PIN for wallet authorization (will be encrypted)",
            "example": "1234",
            "minLength": 4,
            "maxLength": 4
          },
          "walletType": {
            "type": "string",
            "description": "Recipient wallet type",
            "enum": [
              "NORMAL",
              "SALARY",
              "MERCHANT",
              "DISBURSEMENT"
            ],
            "default": "NORMAL"
          },
          "transactionType": {
            "type": "string",
            "description": "Transaction type classification",
            "enum": [
              "B2C",
              "B2B",
              "G2C",
              "B2G",
              "SALARY",
              "LOAN",
              "DIVIDEND",
              "COMMISSION",
              "REFUND",
              "OTHER"
            ],
            "default": "B2C"
          }
        },
        "required": [
          "provider",
          "externalId",
          "payeeMsisdn",
          "amount",
          "reference",
          "pin"
        ]
      },
      "DisbursementResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique disbursement identifier",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "provider": {
            "type": "string",
            "description": "Payment provider that handled the transaction",
            "enum": [
              "AIRTEL",
              "MTN"
            ],
            "example": "AIRTEL"
          },
          "tenantId": {
            "type": "string",
            "description": "Tenant owning this disbursement",
            "example": "tenant-001"
          },
          "externalId": {
            "type": "string",
            "description": "Client-provided idempotency key",
            "example": "order-2024-001-disbursement"
          },
          "payeeMsisdn": {
            "type": "string",
            "description": "Recipient mobile number",
            "example": "0977123456"
          },
          "walletType": {
            "type": "string",
            "description": "Recipient wallet type",
            "enum": [
              "NORMAL",
              "SALARY",
              "MERCHANT",
              "DISBURSEMENT"
            ],
            "example": "NORMAL"
          },
          "amount": {
            "type": "string",
            "description": "Disbursement amount",
            "example": "500.50"
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "ZMW"
          },
          "reference": {
            "type": "string",
            "description": "Business reference for reconciliation",
            "example": "INV-2024-001"
          },
          "transactionType": {
            "type": "string",
            "description": "Transaction type classification",
            "enum": [
              "B2C",
              "B2B",
              "G2C",
              "B2G",
              "SALARY",
              "LOAN",
              "DIVIDEND",
              "COMMISSION",
              "REFUND",
              "OTHER"
            ],
            "example": "B2C"
          },
          "status": {
            "type": "string",
            "description": "Current disbursement status",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SUCCESS",
              "FAILED",
              "TIMEOUT",
              "BOUNCED",
              "REFUNDED",
              "REFUND_PROCESSING",
              "REFUND_FAILED"
            ],
            "example": "SUCCESS"
          },
          "airtelReferenceId": {
            "type": "string",
            "description": "Airtel-generated reference ID",
            "example": "AIRTEL-REF-12345",
            "nullable": true
          },
          "airtelMoneyId": {
            "type": "string",
            "description": "Airtel Money transaction ID",
            "example": "MONEY-12345-67890",
            "nullable": true
          },
          "errorCode": {
            "type": "string",
            "description": "Error code if disbursement failed",
            "example": "INVALID_MSISDN",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "The provided MSISDN is invalid or inactive",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-02-06T10:30:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-02-06T10:31:00Z"
          }
        },
        "required": [
          "id",
          "provider",
          "tenantId",
          "externalId",
          "payeeMsisdn",
          "walletType",
          "amount",
          "currency",
          "reference",
          "transactionType",
          "status",
          "airtelReferenceId",
          "airtelMoneyId",
          "errorCode",
          "errorMessage",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "x-api-key": []
    },
    {
      "x-tenant-id": []
    }
  ]
}